+++ packages/video_player/video_player_platform_interface/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 6.6.1
 
-* Implement screen auto-lock control for video playback.
+* Implements screen auto-lock control for video playback.
 * Updates minimum supported SDK version to Flutter 3.35/Dart 3.9.
 
 ## 6.6.0